+++ backend/audit/templates/audit/upload-report.html
@@ -105,5 +105,4 @@
         </div>
     </div>
     {% include "audit-metadata.html" %}
-    <!-- <script src="{% static 'compiled/js/upload-report-pdf.js' %}"></script> -->
 {% endblock content %}
